refactor#94 fix place ui state && api

+++ data/place/src/main/java/com/teamsolply/solply/place/repository/PlaceRepositoryImpl.kt
@@ -94,9 +94,9 @@ class PlaceRepositoryImpl @Inject constructor(
 
     override suspend fun getPlaces(
         townId: Long,
-        mainTagId: Long?,
-        subTagAIdList: List<Long>?,
-        subTagBIdList: List<Long>?
+        mainTagId: Int?,
+        subTagAIdList: List<Int>?,
+        subTagBIdList: List<Int>?
     ): Result<List<PlaceEntity>> = runCatching {
         placeRemoteDataSource.getPlaces(
             townId = townId,
